Refactor to consolidate where server start happens.
